+++ README.md
@@ -105,18 +105,18 @@ A DatTCP server also has a number of hooks for adding custom behavior when diffe
 
 ## Benchmarking
 
-DatTCP comes with some rake tasks for benchmarking it's performance. These generate report text files that should be used to see if any additions or changes have altered it's previous performance. These can be run by doing the following:
+DatTCP comes with some scripts for benchmarking it's performance. These generate report text files that should be used to see if any additions or changes have altered it's previous performance. These can be run by doing the following:
 
 In a shell, start the server:
 
 ```bash
-bundle exec rake bench:server
+bundle exec ruby bench/server.rb
 ```
 
 This will start the server which will begin listening for requests. The report can then be generated by running this rake task in another shell:
 
 ```bash
-bundle exec rake bench:report
+bundle exec ruby bench/report.rb
 ```
 
 This will both output the results to STDOUT and to a report file. When the server is stopped, it will also write out some statistics on how long it spent processing.
