+++ Rakefile
@@ -1,2 +1,1 @@
 require "bundler/gem_tasks"
-require "bench/tasks"
